\name{kknn-package}
\alias{kknn-package}
\docType{package}
\title{
Weighted k-Nearest Neighbors Classification and Clustering 
}
\description{
Weighted k-Nearest Neighbors Classification, Regression and spectral Clustering

The complete list of functions can be displayed with \code{library(help = kknn)}. 
}
\author{
Klaus Schliep  \cr
Klaus Hechenbichler

Maintainer: Klaus Schliep <klaus.schliep@gmail.com>
}
\references{
Hechenbichler K. and Schliep K.P. (2004)  \emph{Weighted k-Nearest-Neighbor Techniques and Ordinal Classification}, Discussion Paper 399, SFB 386, Ludwig-Maximilians University Munich
(\url{http://www.stat.uni-muenchen.de/sfb386/papers/dsp/paper399.ps})

Hechenbichler K. (2005)  \emph{Ensemble-Techniken und ordinale Klassifikation}, PhD-thesis
}
\keyword{ package }


