% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kknn-package.R
\docType{package}
\name{kknn-package}
\alias{kknn-package}
\title{Weighted k-Nearest Neighbors Classification and Clustering}
\description{
Weighted k-Nearest Neighbors Classification, Regression and spectral
Clustering
}
\details{
The complete list of functions can be displayed with \code{library(help =
kknn)}.
}
\references{
Hechenbichler K. and Schliep K.P. (2004) \emph{Weighted
k-Nearest-Neighbor Techniques and Ordinal Classification}, Discussion Paper
399, SFB 386, Ludwig-Maximilians University Munich
(\doi{10.5282/ubm/epub.1769})

Hechenbichler K. (2005) \emph{Ensemble-Techniken und ordinale
Klassifikation}, PhD-thesis
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/KlausVigo/kknn}
  \item Report bugs at \url{https://github.com/KlausVigo/kknn/issues}
}

}
\author{
Klaus Schliep \cr Klaus Hechenbichler

Maintainer: Klaus Schliep <klaus.schliep@gmail.com>
}
\keyword{internal}
